changed the flaky test again, this time by relying on the jsonstring instead of the order object -- because they are now different references but have the same value. I figure since the strings should be more flexible
